+++ logwise_proc/src/info.rs
@@ -5,18 +5,18 @@ use std::collections::VecDeque;
 
 pub fn info_sync_impl(input: TokenStream) -> TokenStream {
     let mut input: VecDeque<_> = input.into_iter().collect();
-    let lformat_result = lformat_impl(&mut input, "formatter".to_string());
+    let lformat_result = lformat_impl(&mut input, "__logwise_formatter".to_string());
     let src = format!(
         r#"
         #[cfg(debug_assertions)]
         {{
             if logwise::log_enabled!(logwise::Level::Info) {{
-                let mut record = logwise::hidden::info_sync_pre(file!(),line!(),column!());
+                let mut __logwise_record = logwise::hidden::info_sync_pre(file!(),line!(),column!());
 
-                let mut formatter = logwise::hidden::PrivateFormatter::new(&mut record);
+                let mut __logwise_formatter = logwise::hidden::PrivateFormatter::new(&mut __logwise_record);
 
                 {LFORMAT_EXPAND}
-                logwise::hidden::info_sync_post(record);
+                logwise::hidden::info_sync_post(__logwise_record);
             }}
         }}
     "#,
@@ -28,15 +28,15 @@ pub fn info_sync_impl(input: TokenStream) -> TokenStream {
 
 pub fn info_async_impl(input: TokenStream) -> TokenStream {
     let mut input: VecDeque<_> = input.into_iter().collect();
-    let lformat_result = lformat_impl(&mut input, "formatter".to_string());
+    let lformat_result = lformat_impl(&mut input, "__logwise_formatter".to_string());
     let src = format!(
         r#"
         #[cfg(debug_assertions)] {{
             if logwise::log_enabled!(logwise::Level::Info) {{
-                let mut record = logwise::hidden::info_sync_pre(file!(),line!(),column!());
-                let mut formatter = logwise::hidden::PrivateFormatter::new(&mut record);
+                let mut __logwise_record = logwise::hidden::info_sync_pre(file!(),line!(),column!());
+                let mut __logwise_formatter = logwise::hidden::PrivateFormatter::new(&mut __logwise_record);
                 {LFORMAT_EXPAND}
-                logwise::hidden::info_async_post(record).await;
+                logwise::hidden::info_async_post(__logwise_record).await;
             }}
         }}
     "#,