+++ logwise_proc/src/profile.rs
@@ -5,18 +5,18 @@ use std::collections::VecDeque;
 
 pub fn profile_begin_impl(input: TokenStream) -> TokenStream {
     let mut input: VecDeque<_> = input.into_iter().collect();
-    let lformat_result = lformat_impl(&mut input, "formatter".to_string());
+    let lformat_result = lformat_impl(&mut input, "__logwise_formatter".to_string());
     let src = format!(
         r#"
         {{
             if logwise::log_enabled!(logwise::Level::Profile) {{
-                let (id, mut record) = logwise::hidden::profile_begin_pre(file!(),line!(),column!());
-                let mut formatter = logwise::hidden::PrivateFormatter::new(&mut record);
+                let (__logwise_id, mut __logwise_record) = logwise::hidden::profile_begin_pre(file!(),line!(),column!());
+                let mut __logwise_formatter = logwise::hidden::PrivateFormatter::new(&mut __logwise_record);
                 {LFORMAT_EXPAND}
-                logwise::hidden::profile_begin_post(id, record, "{NAME}")
+                logwise::hidden::profile_begin_post(__logwise_id, __logwise_record, "{NAME}")
             }} else {{
-                let (id, record) = logwise::hidden::profile_begin_pre(file!(),line!(),column!());
-                logwise::hidden::profile_begin_post(id, record, "{NAME}")
+                let (__logwise_id, __logwise_record) = logwise::hidden::profile_begin_pre(file!(),line!(),column!());
+                logwise::hidden::profile_begin_post(__logwise_id, __logwise_record, "{NAME}")
             }}
         }}
     "#,
@@ -28,17 +28,17 @@ pub fn profile_begin_impl(input: TokenStream) -> TokenStream {
 
 pub fn profile_sync_impl(input: TokenStream) -> TokenStream {
     let mut input: VecDeque<_> = input.into_iter().collect();
-    let lformat_result = lformat_impl(&mut input, "formatter".to_string());
+    let lformat_result = lformat_impl(&mut input, "__logwise_formatter".to_string());
     let src = format!(
         r#"
         {{
             if logwise::log_enabled!(logwise::Level::Profile) {{
-                let mut record = logwise::hidden::profile_sync_pre(file!(),line!(),column!());
+                let mut __logwise_record = logwise::hidden::profile_sync_pre(file!(),line!(),column!());
 
-                let mut formatter = logwise::hidden::PrivateFormatter::new(&mut record);
+                let mut __logwise_formatter = logwise::hidden::PrivateFormatter::new(&mut __logwise_record);
 
                 {LFORMAT_EXPAND}
-                logwise::hidden::profile_sync_post(record);
+                logwise::hidden::profile_sync_post(__logwise_record);
             }}
         }}
     "#,
@@ -50,17 +50,17 @@ pub fn profile_sync_impl(input: TokenStream) -> TokenStream {
 
 pub fn profile_async_impl(input: TokenStream) -> TokenStream {
     let mut input: VecDeque<_> = input.into_iter().collect();
-    let lformat_result = lformat_impl(&mut input, "formatter".to_string());
+    let lformat_result = lformat_impl(&mut input, "__logwise_formatter".to_string());
     let src = format!(
         r#"
         {{
             if logwise::log_enabled!(logwise::Level::Profile) {{
-                let mut record = logwise::hidden::profile_sync_pre(file!(),line!(),column!());
+                let mut __logwise_record = logwise::hidden::profile_sync_pre(file!(),line!(),column!());
 
-                let mut formatter = logwise::hidden::PrivateFormatter::new(&mut record);
+                let mut __logwise_formatter = logwise::hidden::PrivateFormatter::new(&mut __logwise_record);
 
                 {LFORMAT_EXPAND}
-                logwise::hidden::profile_async_post(record).await;
+                logwise::hidden::profile_async_post(__logwise_record).await;
             }}
         }}
     "#,